feat: move policies page and resources manager dialog to own tabs on the settings page
