+++ frontend/lib/components/Navigation/NavigationRail.vue
@@ -142,36 +142,6 @@
     </nav>
 
     <div class="bottom">
-      <!-- Policies -->
-      <RouterLink
-        v-if="authUser.isLocalAdministrator"
-        to="/policies"
-        custom
-        v-slot="{ href, isActive, navigate }"
-      >
-        <RailButton :href="href" :active="isActive" @click="navigate">
-          <template v-slot:icon>
-            <!-- policies svg normal -->
-            <svg width="24" height="24" fill="none" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
-              <path
-                d="M12 1.999c5.487 0 9.942 4.419 10 9.892a6.064 6.064 0 0 1-1.525-.566 8.486 8.486 0 0 0-.21-1.326h-1.942a1.618 1.618 0 0 0-.648 0h-.769c.012.123.023.246.032.37a7.858 7.858 0 0 1-1.448.974c-.016-.46-.047-.908-.094-1.343H8.604a18.968 18.968 0 0 0 .135 5H12v1.5H9.06c.653 2.414 1.786 4.002 2.94 4.002.276 0 .551-.091.819-.263a6.938 6.938 0 0 0 1.197 1.56c-.652.133-1.326.203-2.016.203-5.524 0-10.002-4.478-10.002-10.001C1.998 6.477 6.476 1.998 12 1.998ZM7.508 16.501H4.785a8.532 8.532 0 0 0 4.095 3.41c-.523-.82-.954-1.846-1.27-3.015l-.102-.395ZM7.093 10H3.735l-.004.017a8.524 8.524 0 0 0-.233 1.984c0 1.056.193 2.067.545 3h3.173a20.301 20.301 0 0 1-.218-3c0-.684.033-1.354.095-2.001Zm1.788-5.91-.023.008A8.531 8.531 0 0 0 4.25 8.5h3.048c.313-1.752.86-3.278 1.583-4.41ZM12 3.499l-.116.005C10.618 3.62 9.396 5.622 8.828 8.5h6.343c-.566-2.87-1.784-4.869-3.045-4.995L12 3.5Zm3.12.59.106.175c.67 1.112 1.178 2.572 1.475 4.237h3.048a8.533 8.533 0 0 0-4.338-4.29l-.291-.121Zm7.38 8.888c-1.907-.172-3.434-1.287-4.115-1.87a.601.601 0 0 0-.77 0c-.682.583-2.21 1.698-4.116 1.87a.538.538 0 0 0-.5.523V17c0 4.223 4.094 5.716 4.873 5.962a.42.42 0 0 0 .255 0c.78-.246 4.872-1.74 4.872-5.962v-3.5a.538.538 0 0 0-.5-.523Z"
-                fill="currentColor"
-              />
-            </svg>
-          </template>
-          <template v-slot:icon-active>
-            <!-- policies svg active -->
-            <svg width="24" height="24" fill="none" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
-              <path
-                d="M8.904 16.5H12v.5c0 1.728.59 3.072 1.383 4.09-.439.584-.912.91-1.383.91-1.197 0-2.401-2.095-3.038-5.205l-.058-.294Zm-.23-1.499H12v-1.5c0-.85.676-1.453 1.409-1.52.773-.07 1.484-.33 2.084-.64a25.619 25.619 0 0 0-.069-1.34H8.575l-.033.488A25.982 25.982 0 0 0 8.674 15Zm13.127-5c.124.611.193 1.243.2 1.89-1.33-.287-2.416-1.073-2.965-1.544a1.576 1.576 0 0 0-.707-.346h3.472ZM3.065 16.5h4.306c.365 2.082.983 3.854 1.792 5.093a10.028 10.028 0 0 1-5.951-4.814l-.147-.279Zm-.867-6.5h4.87a28.205 28.205 0 0 0 .033 4.42l.058.58H2.456a10.048 10.048 0 0 1-.258-5Zm12.746-7.424-.108-.17A10.027 10.027 0 0 1 21.372 8.5H16.78c-.316-2.416-.957-4.492-1.837-5.923ZM9.04 2.444l.122-.037c-.88 1.35-1.535 3.329-1.883 5.654l-.061.44H2.628A10.028 10.028 0 0 1 9.04 2.443ZM12 2.002c1.318 0 2.645 2.542 3.214 6.183l.046.315H8.74c.54-3.809 1.904-6.498 3.26-6.498Zm10.5 10.975c-1.907-.172-3.434-1.287-4.115-1.87a.601.601 0 0 0-.77 0c-.682.583-2.21 1.698-4.116 1.87a.538.538 0 0 0-.5.523V17c0 4.223 4.094 5.716 4.873 5.962a.42.42 0 0 0 .255 0c.78-.246 4.872-1.74 4.872-5.962v-3.5a.538.538 0 0 0-.5-.523Z"
-                fill="currentColor"
-              />
-            </svg>
-          </template>
-          {{ $t('policies.title') }}
-        </RailButton>
-      </RouterLink>
-
       <!-- Settings -->
       <RouterLink to="/settings" custom v-slot="{ href, isActive, navigate }">
         <RailButton :href="href" :active="isActive" @click="navigate">
